all changes and improvements

+++ src/test/java/com/as/common/models/commons/AccountDetails.java
@@ -4,14 +4,16 @@ import com.as.common.models.enums.Currency;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @SuperBuilder
 @Data
 public class AccountDetails {
-    private int id;
+    private long id;
     private String owner;
-    private int balance;
+    private BigDecimal balance;
     private Currency currency;
     private String createdAt;
 }
